+++ src/components/ui/input.jsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { cn } from "@/lib/utils";
 
-const Input = React.forwardRef(({ className, type, ...props }) => {
+const Input = React.forwardRef(({ className, type, ...props }, ref) => {
   return (
     <input
       type={type}
@@ -10,6 +10,7 @@ const Input = React.forwardRef(({ className, type, ...props }) => {
         className,
       )}
       {...props} // Spread any other props
+      ref={ref}
     />
   );
 });
