+++ src/containers/Event/steps/Lineup.jsx
@@ -1,9 +1,28 @@
-import React from "react";
+import TextField from "@/components/TextField";
+import TicketTierAndArtist from "../elements/TicketTierAndArtist";
+import { Calendar } from "lucide-react";
+import { Button } from "@/components/ui/button";
 
-const Lineup = () => {
+const Lineup = ({ register }) => {
   return (
-    <div>
-      <h1>Basic Info</h1>
+    <div className="space-y-5 w-full">
+      {/* <TicketTierAndArtist  /> */}
+      <TextField
+        label="Door Open"
+        required={true}
+        type="datetime-local"
+        placeholder="Choose date and time"
+        Icon={Calendar}
+        register={register}
+      />
+      <div className="flex justify-end gap-3 py-8 mb-4 border-t mt-14 border-grey-light">
+        <Button variant="outline" className="w-40">
+          Cancel
+        </Button>
+        <Button className="w-40" disabled={true}>
+          Continue
+        </Button>
+      </div>
     </div>
   );
 };
